``` chore: Update check workflow and create release action
Refactor the check workflow by introducing a new step to create a release using the actions/create-release@v1 action. This step is triggered on a push event to the master branch. The release is created with the tag name as the GitHub run number and the release name as "Release [run number]". The release is not a draft and not a pre-release.

```

Co-authored-by: Sheila Toledo <sheilatoledoblanco@gmail.com>